Create README.md
This is a To-Do-List/Grocery-List I made for my GF to keep track of the items needed when going grocery shopping. It allows her to ADD, DELETE, and even EDIT this this to the items needed when shopping. I created this with HTML, CSS, and Javascript to give the functionality of adding, deleting, and editing the grocery list to specific likings.